requirements.txt
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.2
django-filter==23.2
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
djangorestframework==3.14.0
drf-yasg==1.21.6
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
openapi-codec==1.3.2
packaging==23.1
Pillow==9.5.0
pytz==2023.3
PyYAML==6.0
requests==2.31.0
simplejson==3.19.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.6.3
uritemplate==4.1.1
urllib3==2.0.3
